backend/requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.38.18
botocore==1.38.18
cached_path==1.7.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cn2an==0.5.22
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dateparser==1.1.8
DateTime==5.5
decorator==5.2.1
Deprecated==1.2.18
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
encodec==0.1.1
eng_to_ipa==0.0.2
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.25.0rc1
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.29.1
gradio_client==1.10.1
groovy==0.1.2
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.9
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
networkx==2.8.8
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.77.0
openai-whisper==20240930
orjson==3.10.18
packaging==25.0
pandas==1.5.3
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
proces==0.1.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pykakasi==2.2.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.50.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rsa==4.9.1
ruff==0.11.10
s3transfer==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.13.3
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.27.4
triton==3.3.0
txtsplit==1.0.0
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.4.0
uvicorn==0.34.2
wasabi==0.10.1
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zope.interface==7.2
